# fuse L2Decay and momentum when param.regularizer is set