--- a/python/paddle/optimizer/momentum.py
+++ b/python/paddle/optimizer/momentum.py
@@ -19,8 +19,6 @@ from ..fluid.framework import Variable, name_scope
 from ..fluid.layer_helper import LayerHelper
 from ..fluid import unique_name
 from ..fluid import layers
-from ..fluid.clip import append_gradient_clip_ops
-from ..fluid.regularizer import append_regularization_ops
 import paddle.fluid as fluid
 from paddle.fluid.regularizer import L2DecayRegularizer
 
@@ -254,77 +252,51 @@ class Momentum(Optimizer):
                 )
             self._add_accumulator(self._velocity_acc_str, p)
 
-    def apply_gradients(self, params_grads):
+    def _create_regularization_of_grad(self, param, grad, regularization=None):
+        """ Create and add backward regularization Operators
+    
+        Function helper of append_regularization_ops.
         """
-        Second part of `minimize`, appending optimization operators for
-        given `params_grads` pairs.
+        # (1) If no gradient or no regularization is specified,  then we don't need to do anything.
+        # (2) If ParamAttr is set to L2Decay, we skip doing regularization here. And then we fused
+        # L2Decay with momentum which can refer to _append_optimize_op below.
+        no_regularization = (not hasattr(param, 'regularizer') or (
+            hasattr(param, 'regularizer') and
+            param.regularizer is None)) and regularization is None
+        param_has_L2Decay = hasattr(param, 'regularizer') and isinstance(
+            param.regularizer, L2DecayRegularizer)
+        if grad is None or no_regularization or param_has_L2Decay:
+            return grad
+        regularization_term = None
+        if hasattr(param, 'regularizer') and param.regularizer is not None:
+            # Add variable for regularization term in grad block
+            regularization_term = param.regularizer(param, grad, grad.block)
+        elif regularization is not None:
+            regularization_term = regularization(param, grad, grad.block)
 
-        Args:
-            params_grads (list): list of (param, grad) pair to do optimization.
+        assert regularization_term is not None
 
-        Returns:
-            list: A list of operators appended to the current program.
+        new_grad = grad
+        if grad.type == core.VarDesc.VarType.SELECTED_ROWS:
+            # FIXME(zcd): If the grad is SELECTED_ROWS, after regularization,
+            # the grad's type and name will be changed. But the gradient's name
+            # is used in ParallelExecutor Reduce mode, so I add a flag for
+            # the new_grad here.
+            new_grad = grad.block.create_var(
+                name=grad.name + core.kNewGradSuffix(),
+                dtype=param.dtype,
+                shape=param.shape,
+                lod_level=param.lod_level,
+                type=core.VarDesc.VarType.LOD_TENSOR)
 
-        Examples:
-            .. code-block:: python
-
-                import paddle
-                import numpy as np
-
-                inp = np.random.uniform(-0.1, 0.1, [10, 10]).astype("float32")
-                linear = paddle.nn.Linear(10, 10)
-                inp = paddle.to_tensor(inp)
-                out = linear(inp)
-                loss = paddle.mean(out)
-                optimizer = paddle.optimizer.Adam(learning_rate=0.1,
-                        parameters=linear.parameters())
-                params_grads = optimizer.backward(loss)
-                optimizer.apply_gradients(params_grads)
-
-        """
-
-        params_grads = sorted(params_grads, key=lambda x: x[0].name)
-
-        # 'optimizer(grad_clip)' or 'set_gradient_clip'
-        if self._grad_clip is not None:
-            params_grads = self._grad_clip(params_grads)
-        else:
-            params_grads = append_gradient_clip_ops(params_grads)
-
-        # Add regularization if any
-        if self.regularization is not None:
-            params_grads = append_regularization_ops(params_grads,
-                                                     self.regularization)
-
-        optimize_ops = self._create_optimization_pass(params_grads)
-        return optimize_ops
-
-    def _apply_optimize(self, loss, startup_program, params_grads):
-        """
-        Second part of `minimize`, appending optimization operators for
-        given `params_grads` pairs.
-        Args:
-            loss (Tensor): loss tensor to run optimizations.
-            startup_program (Program): startup_program for initializing parameters
-                in `parameters`.
-            params_grads (list): list of (param, grad) pair to do optimization.
-        Returns:
-            list: A list of operators appended to the current program.
-        """
+        inputs = {"X": [grad, regularization_term]}
+        outputs = {"Out": [new_grad]}
         if framework.in_dygraph_mode():
-            with framework.program_guard(framework.default_main_program(),
-                                         framework.default_startup_program()):
-                if self._grad_clip is not None:
-                    params_grads = self._grad_clip(params_grads)
-                if self.regularization is not None:
-                    params_grads = append_regularization_ops(
-                        params_grads, self.regularization)
-                optimize_ops = self._create_optimization_pass(params_grads)
+            new_grad = core.ops.sum([grad, regularization_term])
         else:
-            program = loss.block.program
-            with framework.program_guard(program, startup_program):
-                optimize_ops = self.apply_gradients(params_grads)
-        return optimize_ops
+            grad.block.append_op(type='sum', inputs=inputs, outputs=outputs)
+
+        return new_grad
 
     def _append_optimize_op(self, block, param_and_grad):
         assert isinstance(block, framework.Block)
@@ -336,10 +308,15 @@ class Momentum(Optimizer):
         lr = self._create_param_lr(param_and_grad)
 
         param = param_and_grad[0]
-        if hasattr(param, 'regularizer') and isinstance(param.regularizer,
-                                                        L2DecayRegularizer):
-            self._regularization_method = "l2_decay"
-            self._regularization_coeff = param.regularizer._regularization_coeff
+        if hasattr(param, 'regularizer'):
+            # we skip param's l2decay before, so fuse it with momentum here.
+            if isinstance(param.regularizer, L2DecayRegularizer):
+                self._regularization_method = "l2_decay"
+                self._regularization_coeff = param.regularizer._regularization_coeff
+            # the param's regularization has been done before, we avoid do l2decay in momentum.
+            elif param.regularizer is not None:
+                self._regularization_method = ""
+                self._regularization_coeff = 0
 
         if framework.in_dygraph_mode():
             if isinstance(param_and_grad, dict):