… python/paddle/fluid/regularizer.py
… python/paddle/fluid/tests/unittests/test_regularizer.py
… python/paddle/optimizer/optimizer.py